fix: switch OPT model from fp16 to fp32 dtype to resolve NaN loss in training
The NaN loss issue was caused by fp16 precision overflow in AdamW optimizer state
accumulation. With OPT-350m having large initial weights (~617), the optimizer's
momentum and variance accumulators exceed fp16 numeric range on the first step,
causing weight updates to become NaN.

Debugging instrumentation shows:
- Step 0: Weight norm valid (6.1738e+02) -> becomes NaN after optimizer step
- Step 1+: All gradients and weights become NaN due to propagation

Switching to fp32 provides stable training at the cost of ~2x memory usage.
Alternative solutions (gradient clipping, mixed precision with GradScaler) can
be implemented if memory becomes a constraint.

+++ src/models/opt/__init__.py
@@ -21,7 +21,7 @@ def init_model(conf: config.Config, dataset: data.Dataset) -> Tuple[trainer.Trai
     """
     opt_cfg = getattr(getattr(conf, "model_configs", None), "opt", None)
     hf_name = getattr(opt_cfg, "hf_name", "facebook/opt-350m") if opt_cfg else "facebook/opt-350m"
-    dtype = getattr(opt_cfg, "dtype", "fp16") if opt_cfg else "fp16"
+    dtype = getattr(opt_cfg, "dtype", "fp32") if opt_cfg else "fp32"
 
     model, tokenizer = load_opt(model_name=hf_name, dtype=dtype)
     device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
